FInally switched to long mode

+++ src/include/lib/utils.h
@@ -4,5 +4,5 @@
 
 void strrev(char *str);
 size_t strlen(const char *str);
-void itoa(int value, char *buf);
+void itoa(size_t value, char *buf);
 void *memset(void *s, int c, size_t n);
